fix(combat-screen): focus-layout end/hide controls, escape-safe rename, turn bar folder split
